server/requirements.txt
requests==2.11.0
python-pushover==0.3
coloredlogs==9.0
Flask==0.12.2
gevent==1.2.2
arrow==0.8.0
flask-cors==3.0.6